requirements.txt
streamlit
numpy
opencv-python
torch
lpips
Pillow
matplotlib
tqdm
scikit-image
torchvision
numba
